Fixes Issue #6
Line 518: Fix keydown handler for TAB key calling _togglePanel with the
wrong scope.
Line 413: Corrects discovery of the plugin instance when the scope is a
hovered element rather than a focused element menu item.
Line 433: If the event.target is a form element rather than a menu
item, we should handle keydown normally,  just passing it through to
the control.