Issue #7: TypeError: that is null when hovering with mouse and using keyboard
_keyDownHandler is called with a different scope depending on whether
or not the target element is a hovered menu item or a focused menu item.

With a focused menu item, _keyDownHandler is called with the
AccessibleMegaMenu instance as the scope, i.e. this==accessibleMegaMenu
instance and event.target==menuItem.

However, when a menu item is hovered without having keyboard focus, we
attach the keydown.accessible-megamenu listener to the document body
with the hovered item as the scope, i.e. this==menuItem instance and
event.target==document.body.

To correctly identify the plugin instance when a menu item is hovered
without having keyboard focus, we should therefore call
_getPlugin(this).

This also corrects determination of whether the target menu item is
hovered and tabbable, by using the settings.hoverClass rather than
.hover.