(Issue #21) Fix for links don't work with JAWS screenreader
Removed aria-owns from top-level navigation links.

JAWS interprets this as the panel and all its children being part of
the content of the top-level link; when you bring up the list of links
using Insert+F7, the list item for the expanded top-level link includes
all of the text of the panel as well and the links within the panel do
not appear in the list box.

By removing the aria-owns attribute and keeping only the aria-controls
attribute, the links within and expanded panel are treated as
individual links as one would expect.

This patch also remove some redundancy in the selector for adding and
removing events on `:focusable` elements.