Add delay (hover intent) to _mouseOverHandler and settings options
Changed _mouseOverHandler to add delay to check for hover intent. If the mouse is simply passing over the menu en route to other content, it won't trigger unless it hovers for a set duration.  Default is set to 0 milliseconds to match previous settings.

Also made made the delay an option in the config so one can readily choose if/when a delay should happen in regard to mouse Over/Out.