Add openOnMouseover setting, and toggleButtonClass for responsive layout
To normalize behavior of top navigation links on desktop and mobile,
the default behavior will be for the menu to remain closed on hover.
Click should toggle the menu regardless of whether the link has a valid
`href`.

With `openOnMouseover: true`, hover will open the menu as before and
with the menu open, clicking a top-level navigation item with a valid
`href` should open the url.

Add toggleButtonClass to specify a button that should toggle the menu
in a responsive layout on mobile devices.

Incorporate media queries for responsive layout breakpoints.